Remove PreferenceScreen and redesign DetailScreen to include notification preference
